refactor, style: separated groupInfo tabs directory, modify style slightly
