+++ src/pages/detail/DetailPage.tsx
@@ -8,9 +8,12 @@ import { getGroup } from "@/apis/group";
 import GroupProfileDefault from "@/assets/icons/group-profile-default.webp";
 
 import GroupDetailTabs from "./GroupDetailTabs";
-import GroupIntroTab from "./GroupIntroTab";
-import GroupMembersTab from "./GroupMembersTab";
-import GroupNoticesTab from "./GroupNoticesTab";
+
+import GroupMembersTab from "./tabs/members/GroupMembersTab";
+import GroupNoticesTab from "./tabs/notices/GroupNoticesTab";
+import GroupIntroTab from "./tabs/intro/GroupIntroTab";
+import GroupProfile from "./GroupProfile";
+import Loading from "@/components/loading/Loading";
 
 interface GroupDetailPageProps {
   searchParams?: { tab: string; page: string };
@@ -21,47 +24,16 @@ const GroupDetailPage = ({ searchParams }: GroupDetailPageProps) => {
 
   const { data: group } = useSWR(uuid, getGroup);
 
-  const { t } = useTranslation();
-
   const [tab, setTab] = useState("info");
 
   if (!group) {
-    return null; // TODO: loading or error page
+    return <Loading />; // TODO: loading or error page
   }
 
   return (
-    <main className={"flex flex-col items-center"}>
+    <main className={"flex flex-col items-center mt-[20px] md:mt-[48px]"}>
       <div className={"content flex max-w-[800px] flex-col"}>
-        <div className={"flex items-center gap-[25px]"}>
-          <img
-            src={GroupProfileDefault}
-            width={160}
-            height={160}
-            alt={"group default profile"}
-          />
-
-          <div className={"flex flex-col items-start"}>
-            <p className={"text-[34px] font-bold leading-9"}>{group.name}</p>
-
-            <p className={"mt-1 text-greyDark"}>
-              {t("group.memberCount", {
-                count: group.memberCount,
-              })}
-              {" · "}
-              {t("group.noticeCount", {
-                count: 0,
-              })}
-            </p>
-
-            <Button
-              variant="contained"
-              className={"mt-3 rounded-[10px] md:px-6 md:py-2"}
-            >
-              <p>{t("group.favorite")}</p>
-            </Button>
-          </div>
-        </div>
-
+        <GroupProfile group={group} />
         <p
           className={
             "my-6 w-full rounded-2xl bg-greyLight px-5 py-[15px] text-lg text-greyDark"
@@ -69,10 +41,9 @@ const GroupDetailPage = ({ searchParams }: GroupDetailPageProps) => {
         >
           {group.description}
         </p>
-
         <GroupDetailTabs activeTab={tab} setActiveTab={setTab} />
-
-        {tab === "info" && <GroupIntroTab />}
+        {/* disabled intro notion tab since not working */}
+        {/* {tab === "info" && <GroupIntroTab />} */}{" "}
         {tab === "notice" && <GroupNoticesTab searchParams={searchParams} />}
         {tab === "member" && <GroupMembersTab />}
       </div>
