+++ src/pages/detail/tabs/intro/GroupIntroTab.tsx
@@ -7,14 +7,17 @@ import "katex/dist/katex.min.css";
 import "./styles.css";
 
 import useSWR from "swr";
-import { getNotionPage } from "../../apis/notion";
+
+import { getNotionPage } from "@/apis/notion";
 import NotionWrapper from "./NotionWrapper";
 
 interface GroupIntroTabProps {}
 
 const GroupIntroTab = ({}: GroupIntroTabProps) => {
-
-  const { data: recordMap } = useSWR('2024-GIST-Developers-Night-1292632ae49843a79fb3e554c6a926c1', getNotionPage);
+  const { data: recordMap } = useSWR(
+    "2024-GIST-Developers-Night-1292632ae49843a79fb3e554c6a926c1",
+    getNotionPage,
+  );
 
   if (recordMap == null) {
     return <div>Notion page not found</div>;
