+++ src/pages/detail/tabs/intro/styles.css
