+++ src/pages/detail/tabs/notices/GroupNoticesTab.tsx
